requirements_cuda.txt
bokeh==3.4.1
ffmpegio==0.8.3
ffmpegio_core==0.8.3
GitPython==3.1.42
numpy==1.26.4
palettable==3.3.3
Pillow==10.3.0
pytritech==1.0.2
pytz==2024.1
scipy==1.13.0
scikit-image==0.23.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.2
wandb==0.16.3
onnxscript==0.1.0.dev20240612
git+https://github.com/OniDaito/pypam.git@main#egg=pypam
git+https://github.com/OniDaito/sealhits.git@main#egg=sealhits
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/
onnxruntime-gpu
tensorrt==10.1.0